+++ src/app/inventory/d2-stores.ts
@@ -355,6 +355,7 @@ function processCharacter(
   const characterInventory = profileInfo.characterInventories.data?.[characterId]?.items || [];
   const profileInventory = profileInfo.profileInventory.data?.items || [];
   const characterEquipment = profileInfo.characterEquipment.data?.[characterId]?.items || [];
+  const profileRecords = profileInfo.profileRecords?.data; // Not present in the initial load
   const itemComponents = profileInfo.itemComponents;
   const uninstancedItemObjectives =
     getCharacterProgressions(profileInfo, characterId)?.uninstancedItemObjectives || [];
@@ -381,7 +382,8 @@ function processCharacter(
     items,
     itemComponents,
     mergedCollectibles,
-    uninstancedItemObjectives
+    uninstancedItemObjectives,
+    profileRecords
   );
   store.items = processedItems;
   return store;
@@ -398,6 +400,7 @@ function processVault(
   const profileInventory = profileInfo.profileInventory.data
     ? profileInfo.profileInventory.data.items
     : [];
+  const profileRecords = profileInfo.profileRecords?.data; // Not present in the initial load
   const itemComponents = profileInfo.itemComponents;
 
   const store = makeVault();
@@ -417,7 +420,9 @@ function processVault(
     store,
     items,
     itemComponents,
-    mergedCollectibles
+    mergedCollectibles,
+    undefined,
+    profileRecords
   );
   store.items = processedItems;