+++ src/app/search/search-filters/sockets.tsx
@@ -7,7 +7,7 @@ import {
   modTypeTags,
 } from 'app/utils/item-utils';
 import { getSocketsByCategoryHash } from 'app/utils/socket-utils';
-import { DestinyItemSubType } from 'bungie-api-ts/destiny2';
+import { DestinyItemSubType, DestinyRecordState } from 'bungie-api-ts/destiny2';
 import craftingMementos from 'data/d2/crafting-mementos.json';
 import {
   ItemCategoryHashes,
@@ -206,7 +206,7 @@ const socketFilters: FilterDefinition[] = [
     description: tl('Filter.Deepsight'),
     format: ['simple', 'query'],
     destinyVersion: 2,
-    suggestions: ['complete', 'incomplete'],
+    suggestions: ['complete', 'incomplete', 'pattern'],
     filter:
       ({ filterValue }) =>
       (item: DimItem) => {
@@ -218,9 +218,14 @@ const socketFilters: FilterDefinition[] = [
           case 'deepsight':
             return true;
           case 'complete':
-            return item.deepsightInfo.complete;
+            return item.deepsightInfo.attunementObjective.complete;
           case 'incomplete':
-            return !item.deepsightInfo.complete;
+            return !item.deepsightInfo.attunementObjective.complete;
+          case 'pattern':
+            return Boolean(
+              item.patternUnlockRecord &&
+                item.patternUnlockRecord.state & DestinyRecordState.ObjectiveNotCompleted
+            );
         }
       },
   },